+++ client/Packages/com.beamable.server/Editor/UI/Components/ServiceBaseVisualElement/CreateServiceBaseVisualElement.cs
@@ -40,7 +40,7 @@ namespace Beamable.Editor.Microservice.UI.Components
         private Button _moreBtn;
         private Button _cancelBtn;
         private Button _buildDropDownBtn;
-        private BeamableCheckboxVisualElement _checkbox;
+        private LabeledCheckboxVisualElement _checkbox;
         private Button _createBtn;
         private VisualElement _logContainerElement;
         private Label _buildDefaultLabel;
@@ -61,12 +61,11 @@ namespace Beamable.Editor.Microservice.UI.Components
         protected virtual void QueryVisualElements()
         {
             _rootVisualElement = Root.Q<VisualElement>("mainVisualElement");
-            Root.Q("microserviceTitle")?.RemoveFromHierarchy();
             Root.Q("dependentServicesContainer")?.RemoveFromHierarchy();
             _cancelBtn = Root.Q<Button>("cancelBtn");
-            _createBtn = Root.Q<Button>("start");
+            _createBtn = Root.Q<Button>("stopBtn");
             _buildDropDownBtn = Root.Q<Button>("buildDropDown");
-            _checkbox = Root.Q<BeamableCheckboxVisualElement>("checkbox");
+            _checkbox = Root.Q<LabeledCheckboxVisualElement>("checkbox");
             _logContainerElement = Root.Q<VisualElement>("logContainer");
             _nameTextField = Root.Q<TextField>("microserviceNewTitle");
         }
@@ -77,7 +76,7 @@ namespace Beamable.Editor.Microservice.UI.Components
         }
         protected virtual void UpdateVisualElements()
         {
-            _servicesNames = MicroservicesDataModel.Instance.AllServices.Select(x => x.Descriptor.Name).ToList();
+            _servicesNames = MicroservicesDataModel.Instance.AllLocalServices.Select(x => x.Descriptor.Name).ToList();
             RegisterCallback<MouseDownEvent>(HandeMouseDownEvent, TrickleDown.TrickleDown);
             
             _nameTextField.SetValueWithoutNotify(NewServiceName);
@@ -95,6 +94,7 @@ namespace Beamable.Editor.Microservice.UI.Components
             _checkbox.Refresh();
             _checkbox.SetWithoutNotify(false);
             _checkbox.SetEnabled(false);
+            _checkbox.DisableLabel();
 
             _logContainerElement.RemoveFromHierarchy();
             RenameGestureBegin();